change request type to uint64_t

+++ memory-access.c
@@ -22,7 +22,7 @@
 #include "exec/cpu-common.h"
 
 struct request{
-    uint8_t type;      // 0 quit, 1 read, 2 write, ... rest reserved
+    uint64_t type;      // 0 quit, 1 read, 2 write, ... rest reserved
     uint64_t address;  // address to read from OR write to
     uint64_t length;   // number of bytes to read OR write
 };
